move functions to utils files in order to facilitate middleware removal
